Move reporters to .reporter
