+++ newsroom/src/main/java/com/theguardian/newsroom/Reporter.kt
@@ -1,13 +1,15 @@
 package com.theguardian.newsroom
 
 import com.theguardian.newsroom.model.Event
+import com.theguardian.newsroom.reporter.Newsroom
 
 
 abstract class Reporter {
 
     open val sourceName: String = ""
     private var _newsroom: Newsroom? = null
-    private val newsroom: Newsroom get() {
+    private val newsroom: Newsroom
+        get() {
         return _newsroom ?: throw IllegalStateException("This reporters gone rogue, give them a newsroom")
     }
