+++ newsroom/src/main/java/com/theguardian/newsroom/reporter/GoogleAnalyticsReporter.kt
@@ -1,6 +1,7 @@
-package com.theguardian.newsroom
+package com.theguardian.newsroom.reporter
 
 import android.util.Log
+import com.theguardian.newsroom.Reporter
 import com.theguardian.newsroom.model.Event
 import io.reactivex.Observable
 import io.reactivex.disposables.Disposable
